pass the chip dir object to the exported gpio dir's constructor
We need to store the fs entry representing the GPIO chip's directory in
every exported GPIO object. For now we only use the chip's name but
we'll soon want to know the GPIO's parent's indentity in order to be
able to remove still exported GPIOs when the chip that exposed them is
removed.

Signed-off-by: Bartosz Golaszewski <bartosz.golaszewski@linaro.org>