exit on errors in the udev event handler
If we encounter an error in the udev thread, it will exit and the
program will no longer process GPIO chip events. In order to make the
user aware of it, let's exit the entire process. This will then result
in ENOTCONN on any access to the FUSE filesystem that no longer has a
backing process.

Signed-off-by: Bartosz Golaszewski <bartosz.golaszewski@linaro.org>